--- a/gatsby-config.js
+++ b/gatsby-config.js
@@ -37,7 +37,6 @@ module.exports = {
       },
     },
     `gatsby-plugin-sass`,
-    `gatsby-transformer-json`,
     `gatsby-transformer-sharp`,
     `gatsby-plugin-sharp`,
     {