--- a/src/functions/build/index.js
+++ b/src/functions/build/index.js
@@ -2,10 +2,12 @@ import onCreateNode from './onCreateNode';
 import sourceNodes from './sourceNodes';
 import createResolvers from './createResolvers';
 import createPages from './createPages';
+import onCreateWebpackConfig from './onCreateWebpackConfig';
 
 export {
   onCreateNode,
   sourceNodes,
   createResolvers,
   createPages,
+  onCreateWebpackConfig,
 };