--- a/src/organisms/snippetList/snippetList.test.jsx
+++ b/src/organisms/snippetList/snippetList.test.jsx
@@ -16,15 +16,10 @@ describe('<SnippetList />', () => {
   const snippetList = [
     {
       title: 'compose',
-      language: {
-        long: 'JavaScript',
-        short: 'js',
-      },
+      language: 'JavaScript',
       primaryTag: 'function',
       expertise: 'intermediate',
-      html: {
-        description: '<p>Performs right-to-left function composition.</p>',
-      },
+      description: '<p>Performs right-to-left function composition.</p>',
       url: 'snippets/compose',
     },
   ];