--- a/src/queries/getSearchIndex.js
+++ b/src/queries/getSearchIndex.js
@@ -15,7 +15,6 @@ export default `
         }
         expertise
         language {
-          short
           long
         }
         searchTokens