--- a/src/templates/listingPage/listingPage.test.jsx
+++ b/src/templates/listingPage/listingPage.test.jsx
@@ -22,15 +22,10 @@ describe('<ListingPage />', () => {
   const snippetList = [
     {
       title: 'compose',
-      language: {
-        long: 'JavaScript',
-        short: 'js',
-      },
+      language: 'JavaScript',
       primaryTag: 'function',
       expertise: 'intermediate',
-      html: {
-        description: '<p>Performs right-to-left function composition.</p>',
-      },
+      description: '<p>Performs right-to-left function composition.</p>',
       url: 'snippets/compose',
     },
   ];