There were multiple issues slowing down both builds and the website itself. Namely:

- `gatsby-transformer-json` plugin was redundant.
- Source maps were being generated in production.
- There was a bug with `</p>` being cropped in some snippet descriptions, which would cause some slowdowns or even errors.
- Preview card data had a maximum depth of 2, which would slow performance and create additional data that was not necessary.

The performance and size metrics before and after the changes are as follows:

|Metric|Value before|Value after|Gain|
|---|---|---|---|
|Directory size| 70.6MB | 63.5MB| 9% |
|Build speed (Gatsby/Production)|59s|47s|8%|
|Performance rating (Lighthouse)|94|98|4%|
|Speed Index| 2.6s|2.0s|30%|
|Main thread work|3.0s|2.6s|15%|